adding sqlitel1test
